+++ src/modules/sbstudio/plugin/panels/formations.py
@@ -6,6 +6,7 @@ from sbstudio.plugin.operators import (
     CreateFormationOperator,
     CreateTakeoffGridOperator,
     DeselectFormationOperator,
+    FormationHoverTestOperator,
     GetFormationStatisticsOperator,
     LandOperator,
     RemoveFormationOperator,
@@ -57,23 +58,26 @@ class FormationsPanel(Panel):
         layout.operator(CreateTakeoffGridOperator.bl_idname, icon="ADD")
 
         row = layout.row(align=True)
+        row.operator(
+            SingleDroneHoverTestOperator.bl_idname,
+            text="Single Hover",
+            icon="SORTTIME",
+        )
+        row.operator(
+            SingleBoxHoverTestOperator.bl_idname,
+            text="Box Hover",
+            icon="SNAP_FACE_CENTER",
+        )
+        row.operator(
+            FormationHoverTestOperator.bl_idname,
+            text="Formation Hover",
+            icon="OUTLINER_OB_GROUP_INSTANCE",
+        )
+
+        row = layout.row(align=True)
         row.operator(TakeoffOperator.bl_idname, text="Takeoff", icon="TRIA_UP_BAR")
         row.operator(ReturnToHomeOperator.bl_idname, text="RTH", icon="HOME")
         row.operator(LandOperator.bl_idname, text="Land", icon="TRIA_DOWN_BAR")
-
-        row = layout.row(align=True)
-        row.operator(
-            SingleDroneHoverTestOperator.bl_idname,
-            text="Single Drone Hover Test",
-            icon="SORTTIME",
-        )
-
-        row = layout.row(align=True)
-        row.operator(
-            SingleBoxHoverTestOperator.bl_idname,
-            text="Single Box Hover Test",
-            icon="SNAP_FACE_CENTER",
-        )
 
         layout.separator()